+++ src/Sentry/SentryStackTrace.cs
@@ -11,7 +11,7 @@ namespace Sentry;
 /// Frames should be sorted from oldest to newest.
 /// </remarks>
 /// <see href="https://develop.sentry.dev/sdk/event-payloads/stacktrace/"/>
-public sealed class SentryStackTrace : IJsonSerializable
+public class SentryStackTrace : IJsonSerializable
 {
     internal IList<SentryStackFrame>? InternalFrames { get; private set; }