+++ test/Sentry.NLog.Tests/IntegrationTests.verify.cs
@@ -55,6 +55,7 @@ public class IntegrationTests
         LogManager.Flush();
 
         return Verify(transport.Envelopes)
+            .UniqueForRuntimeAndVersion()
             .IgnoreStandardSentryMembers();
     }