+++ test/Sentry.Tests/HubTests.verify.cs
@@ -1,5 +1,3 @@
-using Sentry.Testing;
-
 namespace Sentry.Tests;
 
 [UsesVerify]
@@ -40,8 +38,16 @@ public partial class HubTests
         });
 
         await Verify(worker.Envelopes)
+            .UniqueForRuntimeAndVersion()
             .IgnoreStandardSentryMembers()
             .IgnoreMember("Stacktrace")
-            .IgnoreMember<SentryThread>(_ => _.Name);
+            .IgnoreMember<SentryThread>(_ => _.Name)
+            .IgnoreInstance<DebugImage>(_ =>
+                _.DebugFile != null && (
+                    _.DebugFile.Contains("Xunit.SkippableFact") ||
+                    _.DebugFile.Contains("xunit.runner") ||
+                    _.DebugFile.Contains("Microsoft.TestPlatform")
+                )
+            );
     }
 }